+++ src/constants/inputs.ts
