+++ src/index.ts
@@ -1,6 +1,7 @@
+require('babel-core/register');
+
 import { renderDOM, registerComponent } from './core';
 import './styles/style.scss';
-
 import Button from './components/Button/Button';
 import Link from './components/Link/Link';
 import Input from './components/Input/Input';
@@ -14,7 +15,7 @@ import Popup from './components/Popup/Popup';
 
 import StartPage from './pages/start/start';
 import SignupPage from './pages/signup/signup';
-import { inputs } from './data/inputs';
+import { inputs } from './constants/inputs';
 import Main from './pages/main/main';
 import { chats } from './data/chats';
 import { userData } from './data/userData';
@@ -22,8 +23,6 @@ import Profile from './pages/profile/profile';
 import changeUserData from './pages/changeUserData/changeUserData';
 import changeUserPassword from './pages/changeUserPassword/changeUserPassword';
 import changeUserAvatar from './pages/changeUserAvatar/changeUserAvatar';
-
-require('babel-core/register');
 
 registerComponent(Button);
 registerComponent(Link);
